exercise-tracker/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.0.5
newsapi-python==0.2.7
numpy==2.1.0
outcome==1.3.0.post0
pandas==2.2.2
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
twilio==9.2.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
yarl==1.9.6
